LayoutTests/media/wireless-playback-media-player/current-time-updates-after-rate-change-expected.txt


Test that currentTime advances at the correct rate after a playback rate change.

RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
RUN(video.src = findMediaFile('video', '../content/test'))
RUN(route.playbackRate = 1)
RUN(route.timeRange = { start: 0, duration: 60 })
RUN(video.play())
RUN(route.ready = true)
EVENT(canplaythrough)
RUN(route.currentPlaybackPosition = 10)
EVENT(timeupdate)
RUN(video.playbackRate = 2)
EXPECTED (video.currentTime > 14) OK
EXPECTED (video.playbackRate == '2') OK
END OF TEST
